+++ app/(auth)/sign-up.jsx
